add font size settings and display settings
