+++ src/renderer/src/components/settings/CommonSettings.vue
@@ -1,25 +1,6 @@
 <template>
   <ScrollArea class="w-full h-full p-2">
     <div class="w-full h-full flex flex-col gap-1.5">
-      <div class="flex flex-row p-2 items-center gap-2 px-2">
-        <span class="flex flex-row items-center gap-2 flex-grow w-full">
-          <Icon icon="lucide:languages" class="w-4 h-4 text-muted-foreground" />
-          <span class="text-sm font-medium">{{ t('settings.common.language') }}</span>
-        </span>
-        <!-- 语言选择 -->
-        <div class="flex-shrink-0 min-w-64 max-w-96">
-          <Select v-model="selectedLanguage" class="">
-            <SelectTrigger>
-              <SelectValue :placeholder="t('settings.common.languageSelect')" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem v-for="lang in languageOptions" :key="lang.value" :value="lang.value">
-                {{ lang.label }}
-              </SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-      </div>
       <!-- 搜索引擎选择 -->
       <div class="flex flex-row p-2 items-center gap-2 px-2">
         <span class="flex flex-row items-center gap-2 flex-grow w-full">
@@ -136,20 +117,6 @@
           {{ t('settings.common.invalidProxyUrl') }}
         </div>
       </div>
-      <!-- artifacts效果开关 -->
-      <!-- <div class="flex flex-row p-2 items-center gap-2 px-2">
-        <span class="flex flex-row items-center gap-2 flex-grow w-full">
-          <Icon icon="lucide:sparkles" class="w-4 h-4 text-muted-foreground" />
-          <span class="text-sm font-medium">Artifacts</span>
-        </span>
-        <div class="flex-shrink-0">
-          <Switch
-            id="artifacts-effect-switch"
-            :checked="artifactsEffectEnabled"
-            @update:checked="(val) => settingsStore.setArtifactsEffectEnabled(Boolean(val))"
-          />
-        </div>
-      </div> -->
       <!-- 搜索预览开关 -->
       <div class="flex flex-row p-2 items-center gap-2 px-2">
         <span class="flex flex-row items-center gap-2 flex-grow w-full">
@@ -161,22 +128,6 @@
             id="search-preview-switch"
             :checked="searchPreviewEnabled"
             @update:checked="handleSearchPreviewChange"
-          />
-        </div>
-      </div>
-      <!-- 投屏保护开关 -->
-      <div class="flex flex-row p-2 items-center gap-2 px-2">
-        <span class="flex flex-row items-center gap-2 flex-grow w-full">
-          <Icon icon="lucide:monitor" class="w-4 h-4 text-muted-foreground" />
-          <span class="text-sm font-medium">{{
-            t('settings.common.contentProtection') || '投屏保护'
-          }}</span>
-        </span>
-        <div class="flex-shrink-0">
-          <Switch
-            id="content-protection-switch"
-            :checked="contentProtectionEnabled"
-            @update:checked="handleContentProtectionChange"
           />
         </div>
       </div>
@@ -328,39 +279,6 @@
     </DialogContent>
   </Dialog>
 
-  <!-- 投屏保护切换确认对话框 -->
-  <Dialog v-model:open="isContentProtectionDialogOpen">
-    <DialogContent>
-      <DialogHeader>
-        <DialogTitle>{{
-          t('settings.common.contentProtectionDialogTitle') || '确认切换投屏保护'
-        }}</DialogTitle>
-        <DialogDescription>
-          <template v-if="newContentProtectionValue">
-            {{ t('settings.common.contentProtectionEnableDesc') }}
-          </template>
-          <template v-else>
-            {{ t('settings.common.contentProtectionDisableDesc') }}
-          </template>
-          <div class="mt-2 font-medium">
-            {{ t('settings.common.contentProtectionRestartNotice') }}
-          </div>
-        </DialogDescription>
-      </DialogHeader>
-      <DialogFooter>
-        <Button variant="outline" @click="cancelContentProtectionChange">
-          {{ t('dialog.cancel') }}
-        </Button>
-        <Button
-          :variant="newContentProtectionValue ? 'default' : 'destructive'"
-          @click="confirmContentProtectionChange"
-        >
-          {{ t('dialog.confirm') }}
-        </Button>
-      </DialogFooter>
-    </DialogContent>
-  </Dialog>
-
   <!-- 测试搜索引擎确认对话框 -->
   <Dialog v-model:open="isTestSearchEngineDialogOpen">
     <DialogContent>
@@ -424,7 +342,6 @@ const configPresenter = usePresenter('configPresenter')
 const settingsStore = useSettingsStore()
 const { t } = useI18n()
 
-const selectedLanguage = ref('system')
 const selectedSearchEngine = ref(settingsStore.activeSearchEngine?.id ?? 'google')
 const selectedSearchModel = computed(() => settingsStore.searchAssistantModel)
 
@@ -515,18 +432,6 @@ const addCustomSearchEngine = async () => {
 
 let proxyUrlDebounceTimer: number | null = null
 
-const languageOptions = [
-  { value: 'system', label: '跟随系统' },
-  { value: 'zh-CN', label: '简体中文' },
-  { value: 'en-US', label: 'English (US)' },
-  { value: 'zh-TW', label: '繁體中文（台灣）' },
-  { value: 'zh-HK', label: '繁體中文（香港）' },
-  { value: 'ko-KR', label: '한국어' },
-  { value: 'ru-RU', label: 'Русский' },
-  { value: 'ja-JP', label: '日本語' },
-  { value: 'fr-FR', label: 'Français' }
-]
-
 const proxyModes = [
   { value: 'system', label: t('settings.common.proxyModeSystem') },
   { value: 'none', label: t('settings.common.proxyModeNone') },
@@ -550,19 +455,7 @@ const validateProxyUrl = () => {
   }
 }
 
-// const artifactsEffectEnabled = computed({
-//   get: () => {
-//     console.log('获取artifactsEffectEnabled值:', settingsStore.artifactsEffectEnabled)
-//     return settingsStore.artifactsEffectEnabled
-//   },
-//   set: (value) => {
-//     console.log('设置artifactsEffectEnabled值:', value)
-//     settingsStore.setArtifactsEffectEnabled(value)
-//   }
-// })
-
 onMounted(async () => {
-  selectedLanguage.value = settingsStore.language
   selectedSearchEngine.value = settingsStore.activeSearchEngine?.id ?? 'google'
 
   selectedProxyMode.value = await configPresenter.getProxyMode()
@@ -570,11 +463,6 @@ onMounted(async () => {
   if (selectedProxyMode.value === 'custom' && customProxyUrl.value) {
     validateProxyUrl()
   }
-})
-
-watch(selectedLanguage, async (newValue) => {
-  console.log('selectedLanguage', newValue)
-  await settingsStore.updateLanguage(newValue)
 })
 
 watch(selectedSearchEngine, async (newValue) => {
@@ -692,16 +580,6 @@ const searchPreviewEnabled = computed({
   }
 })
 
-// 投屏保护开关
-const contentProtectionEnabled = computed({
-  get: () => {
-    return settingsStore.contentProtectionEnabled
-  },
-  set: (value) => {
-    settingsStore.setContentProtectionEnabled(value)
-  }
-})
-
 // 日志开关
 const loggingEnabled = computed({
   get: () => {
@@ -716,14 +594,6 @@ const loggingEnabled = computed({
 const handleSearchPreviewChange = (value: boolean) => {
   console.log('切换搜索预览状态:', value)
   settingsStore.setSearchPreviewEnabled(value)
-}
-
-// 处理投屏保护状态变更
-const handleContentProtectionChange = (value: boolean) => {
-  console.log('准备切换投屏保护状态:', value)
-  // 显示确认对话框
-  newContentProtectionValue.value = value
-  isContentProtectionDialogOpen.value = true
 }
 
 // 日志开关相关
@@ -749,19 +619,6 @@ const confirmLoggingChange = () => {
 
 const openLogFolder = () => {
   configPresenter.openLoggingFolder()
-}
-
-// 投屏保护切换确认对话框
-const isContentProtectionDialogOpen = ref(false)
-const newContentProtectionValue = ref(false)
-
-const cancelContentProtectionChange = () => {
-  isContentProtectionDialogOpen.value = false
-}
-
-const confirmContentProtectionChange = () => {
-  settingsStore.setContentProtectionEnabled(newContentProtectionValue.value)
-  isContentProtectionDialogOpen.value = false
 }
 
 // 测试搜索引擎相关
