+++ src/renderer/src/i18n/en-US/routes.json
@@ -7,5 +7,6 @@
   "settings-mcp": "MCP Settings",
   "settings-database": "Data Settings",
   "settings-about": "About",
-  "settings-shortcut": "Shortcuts"
+  "settings-shortcut": "Shortcuts",
+  "settings-display": "Display settings"
 }
