+++ src/renderer/src/i18n/en-US/settings.json
@@ -291,5 +291,13 @@
     "checkUpdate": "Check Update",
     "checking": "Checking...",
     "latestVersion": "Latest Version"
+  },
+  "display": {
+    "fontSize": "Text size",
+    "text-2xl": "Extremely large",
+    "text-base": "Small",
+    "text-lg": "default",
+    "text-sm": "Extremely small",
+    "text-xl": "big"
   }
 }
