+++ src/renderer/src/i18n/fr-FR/common.json
@@ -54,5 +54,6 @@
   "searchEngineSelect": "Sélectionner un moteur de recherche",
   "searchPreview": "Aperçu de la recherche",
   "selectModel": "Sélectionner un modèle",
-  "title": "Paramètres généraux"
+  "title": "Paramètres généraux",
+  "languageSystem": "Suivez le système"
 }
