+++ src/renderer/src/i18n/fr-FR/dialog.json
@@ -21,6 +21,9 @@
     "confirm": "Nettoyer"
   },
   "fork": {
-    "tag": "Branches"
+    "tag": "Branches",
+    "confirm": "Créer une branche",
+    "description": "Copiez le message du premier message au message actuellement sélectionné dans une nouvelle session où vous pouvez continuer la conversation.",
+    "title": "Créer une branche de session"
   }
 }
