+++ src/renderer/src/i18n/fr-FR/routes.json
@@ -7,5 +7,6 @@
   "settings-mcp": "Paramètres MCP",
   "settings-database": "Paramètres des données",
   "settings-about": "À propos",
-  "settings-shortcut": "Raccourcis"
+  "settings-shortcut": "Raccourcis",
+  "settings-display": "Paramètres d'affichage"
 }
