+++ src/renderer/src/i18n/ja-JP/dialog.json
@@ -21,6 +21,9 @@
     "title": "セッションメッセージをクリアします"
   },
   "fork": {
-    "tag": "枝"
+    "tag": "枝",
+    "confirm": "ブランチを作成します",
+    "description": "最初のメッセージから現在選択されているメッセージにメッセージをコピーして、会話を続けることができます。",
+    "title": "セッションブランチを作成します"
   }
 }
