+++ src/renderer/src/i18n/ja-JP/routes.json
@@ -7,5 +7,6 @@
   "settings-about": "について",
   "settings-database": "データ設定",
   "settings-mcp": "MCP設定",
-  "settings-shortcut": "ショートカット"
+  "settings-shortcut": "ショートカット",
+  "settings-display": "設定を表示します"
 }
