+++ src/renderer/src/i18n/ko-KR/dialog.json
@@ -21,6 +21,9 @@
     "confirm": "정리"
   },
   "fork": {
-    "tag": "가지"
+    "tag": "가지",
+    "confirm": "지점을 만듭니다",
+    "description": "첫 번째 메시지의 메시지를 현재 선택한 메시지로 대화를 계속할 수있는 새 세션으로 복사하십시오.",
+    "title": "세션 브랜치를 만듭니다"
   }
 }
