+++ src/renderer/src/i18n/ko-KR/routes.json
@@ -7,5 +7,6 @@
   "settings-about": "정보",
   "settings-database": "데이터 설정",
   "settings-mcp": "MCP 설정",
-  "settings-shortcut": "단축키"
+  "settings-shortcut": "단축키",
+  "settings-display": "디스플레이 설정"
 }
