+++ src/renderer/src/i18n/ko-KR/settings.json
@@ -291,5 +291,13 @@
     "checkUpdate": "업데이트 확인",
     "checking": "확인 중...",
     "latestVersion": "최신 버전"
+  },
+  "display": {
+    "fontSize": "텍스트 크기",
+    "text-2xl": "매우 큽니다",
+    "text-base": "작은",
+    "text-lg": "기본",
+    "text-sm": "매우 작습니다",
+    "text-xl": "큰"
   }
 }
