+++ src/renderer/src/i18n/ru-RU/common.json
@@ -54,5 +54,6 @@
   "searchEngineSelect": "Выберите поисковую систему",
   "searchPreview": "Поиск предварительного просмотра",
   "selectModel": "Выберите модель",
-  "title": "Общие настройки"
+  "title": "Общие настройки",
+  "languageSystem": "Следуйте системе"
 }
