+++ src/renderer/src/i18n/ru-RU/dialog.json
@@ -21,6 +21,9 @@
     "title": "Очистить сообщения сеанса"
   },
   "fork": {
-    "tag": "Ветви"
+    "tag": "Ветви",
+    "confirm": "Создать ветвь",
+    "description": "Скопируйте сообщение из первого сообщения в выбранное в настоящее время сообщение в новом сеансе, где вы можете продолжить разговор.",
+    "title": "Создать филиал сеанса"
   }
 }
