+++ src/renderer/src/i18n/ru-RU/routes.json
@@ -7,5 +7,6 @@
   "settings-about": "О программе",
   "settings-database": "Настройки данных",
   "settings-shortcut": "Горячие клавиши",
-  "settings-mcp": "Настройки MCP"
+  "settings-mcp": "Настройки MCP",
+  "settings-display": "Настройки отображения"
 }
