+++ src/renderer/src/i18n/ru-RU/settings.json
@@ -291,5 +291,13 @@
     "checkUpdate": "Проверить обновления",
     "checking": "Проверка...",
     "latestVersion": "Последняя версия"
+  },
+  "display": {
+    "fontSize": "Размер текста",
+    "text-2xl": "Чрезвычайно большой",
+    "text-base": "Маленький",
+    "text-lg": "по умолчанию",
+    "text-sm": "Чрезвычайно маленький",
+    "text-xl": "большой"
   }
 }
