+++ src/renderer/src/i18n/zh-CN/routes.json
@@ -7,5 +7,6 @@
   "settings-mcp": "MCP设置",
   "settings-database": "数据设置",
   "settings-about": "关于",
-  "settings-shortcut": "快捷键"
+  "settings-shortcut": "快捷键",
+  "settings-display": "显示设置"
 }
