+++ src/renderer/src/i18n/zh-CN/settings.json
@@ -291,5 +291,13 @@
     "checkUpdate": "检查更新",
     "checking": "检查中...",
     "latestVersion": "已是最新版本"
+  },
+  "display": {
+    "fontSize": "文字大小",
+    "text-sm": "极小",
+    "text-base": "小",
+    "text-lg": "默认",
+    "text-xl": "大",
+    "text-2xl": "特大"
   }
 }
