+++ src/renderer/src/i18n/zh-HK/routes.json
@@ -7,5 +7,6 @@
   "settings-about": "關於",
   "settings-database": "數據設置",
   "settings-shortcut": "快捷鍵",
-  "settings-mcp": "MCP設置"
+  "settings-mcp": "MCP設置",
+  "settings-display": "顯示設置"
 }
