+++ src/renderer/src/i18n/zh-TW/dialog.json
@@ -21,6 +21,9 @@
     "title": "錯誤"
   },
   "fork": {
-    "tag": "分支"
+    "tag": "分支",
+    "confirm": "創建分支",
+    "description": "將從第一條消息到當前選中的消息複製到一個新會話中，您可以在新會話中繼續對話。",
+    "title": "創建會話分支"
   }
 }
