+++ src/renderer/src/i18n/zh-TW/routes.json
@@ -7,5 +7,6 @@
   "settings-provider": "大型語言模型設定",
   "settings-mcp": "MCP設定",
   "settings-about": "關於",
-  "settings-shortcut": "快捷鍵"
+  "settings-shortcut": "快捷鍵",
+  "settings-display": "顯示設置"
 }
